backend1/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14,<4.0
djangorestframework-simplejwt>=5.2,<6.0
asgiref>=3.11.0
python-dotenv>=1.2.1
sqlparse>=0.5.5
djangorestframework-stubs[compatible-mypy]
django-stubs[compatible-mypy]
requests>=2.31,<3.0
